w3c-reports/12092014/mac-wk2/runned-tests/user-timing/idlharness-expected.txt

User Timing IDL tests

partial interface Performance {
    void mark(DOMString markName);
    void clearMarks(optional  DOMString markName);

    void measure(DOMString measureName, optional DOMString startMark, optional DOMString endMark);
    void clearMeasures(optional DOMString measureName);
};

interface PerformanceMark : PerformanceEntry {
};

interface PerformanceMeasure : PerformanceEntry {
};
       

FAIL Performance interface: operation mark(DOMString) assert_own_property: window does not have own property "Performance" expected property "Performance" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1284:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: operation clearMarks(DOMString) assert_own_property: window does not have own property "Performance" expected property "Performance" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1284:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: operation measure(DOMString,DOMString,DOMString) assert_own_property: window does not have own property "Performance" expected property "Performance" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1284:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: operation clearMeasures(DOMString) assert_own_property: window does not have own property "Performance" expected property "Performance" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1284:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance must be primary interface of window.performance assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1526:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_primary_interface_of@http://localhost:8800/resources/idlharness.js:1523:13
test_object@http://localhost:8800/resources/idlharness.js:1495:35
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Stringification of window.performance assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1548:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_primary_interface_of@http://localhost:8800/resources/idlharness.js:1545:9
test_object@http://localhost:8800/resources/idlharness.js:1495:35
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: window.performance must inherit property "mark" with the proper type (0) assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1584:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: calling mark(DOMString) on window.performance with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1625:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: window.performance must inherit property "clearMarks" with the proper type (1) assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1584:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: calling clearMarks(DOMString) on window.performance with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1625:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: window.performance must inherit property "measure" with the proper type (2) assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1584:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: calling measure(DOMString,DOMString,DOMString) on window.performance with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1625:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: window.performance must inherit property "clearMeasures" with the proper type (3) assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1584:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL Performance interface: calling clearMeasures(DOMString) on window.performance with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1625:30
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMark interface: existence and properties of interface object assert_own_property: window does not have own property "PerformanceMark" expected property "PerformanceMark" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:966:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMark interface object length assert_own_property: window does not have own property "PerformanceMark" expected property "PerformanceMark" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1032:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1028:13
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMark interface: existence and properties of interface prototype object assert_own_property: window does not have own property "PerformanceMark" expected property "PerformanceMark" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1073:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMark interface: existence and properties of interface prototype object's "constructor" property assert_own_property: window does not have own property "PerformanceMark" expected property "PerformanceMark" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1161:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMeasure interface: existence and properties of interface object assert_own_property: window does not have own property "PerformanceMeasure" expected property "PerformanceMeasure" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:966:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMeasure interface object length assert_own_property: window does not have own property "PerformanceMeasure" expected property "PerformanceMeasure" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1032:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1028:13
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMeasure interface: existence and properties of interface prototype object assert_own_property: window does not have own property "PerformanceMeasure" expected property "PerformanceMeasure" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1073:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
FAIL PerformanceMeasure interface: existence and properties of interface prototype object's "constructor" property assert_own_property: window does not have own property "PerformanceMeasure" expected property "PerformanceMeasure" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1161:28
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/user-timing/idlharness.html:54:17
global code@http://localhost:8800/user-timing/idlharness.html:55:3)
